refactor(entities): separate Cars.js into /entities/
`BaseCar` is now `Car`
`BaseTruck` is now `Truck`

`Car` inherits from `Entity`

`chassisBody` is now `body`

`import * as Cars from '../entities/Vehicles.js'`
